@@ -4,9 +4,9 @@
   "scripts": {
     "ng": "ng",
     "start": "ng serve dev-app --open",
-    "build": "ng build my-project-name --prod",
-    "build:app": "npm run symlink:copy -- --no-watch && ng build dev-app --prod",
-    "symlink:copy": "symlink copy --angular --packages @my-company-name/my-project-name --prod",
+    "build": "ng build my-project-name --configuration production",
+    "build:app": "npm run symlink:copy -- --no-watch && ng build dev-app --configuration production",
+    "symlink:copy": "symlink copy --angular --packages @my-company-name/my-project-name --configuration production",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e",
@@ -23,34 +23,35 @@
     "@abp/ng.tenant-management": "~4.4.2",
     "@abp/ng.theme.basic": "~4.4.2",
     "@abp/ng.theme.shared": "~4.4.2",
-    "@angular/animations": "~12.0.0",
-    "@angular/common": "~12.0.0",
-    "@angular/compiler": "~12.0.0",
-    "@angular/core": "~12.0.0",
-    "@angular/forms": "~12.0.0",
-    "@angular/platform-browser": "~12.0.0",
-    "@angular/platform-browser-dynamic": "~12.0.0",
-    "@angular/router": "~12.0.0",
+    "@angular/animations": "~12.2.6",
+    "@angular/common": "~12.2.6",
+    "@angular/compiler": "~12.2.6",
+    "@angular/core": "~12.2.6",
+    "@angular/forms": "~12.2.6",
+    "@angular/localize": "~12.2.6",
+    "@angular/platform-browser": "~12.2.6",
+    "@angular/platform-browser-dynamic": "~12.2.6",
+    "@angular/router": "~12.2.6",
     "rxjs": "~6.6.0",
     "tslib": "^2.1.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
     "@abp/ng.schematics": "~4.4.2",
-    "symlink-manager": "^1.5.0",
-    "@angular-devkit/build-angular": "~12.0.4",
-    "@angular-eslint/builder": "12.1.0",
-    "@angular-eslint/eslint-plugin": "12.1.0",
-    "@angular-eslint/eslint-plugin-template": "12.1.0",
-    "@angular-eslint/schematics": "12.1.0",
-    "@angular-eslint/template-parser": "12.1.0",
-    "@angular/cli": "~12.0.0",
-    "@angular/compiler-cli": "~12.0.0",
+    "@angular-devkit/build-angular": "~12.2.6",
+    "@angular-eslint/builder": "~12.5.0",
+    "@angular-eslint/eslint-plugin": "~12.5.0",
+    "@angular-eslint/eslint-plugin-template": "~12.5.0",
+    "@angular-eslint/schematics": "~12.5.0",
+    "@angular-eslint/template-parser": "~12.5.0",
+    "@angular/cli": "~12.2.6",
+    "@angular/compiler-cli": "~12.2.6",
+    "@angular/language-service": "~12.2.6",
     "@types/jasmine": "~3.6.0",
     "@types/node": "^12.11.1",
-    "@typescript-eslint/eslint-plugin": "4.23.0",
-    "@typescript-eslint/parser": "4.23.0",
-    "eslint": "^7.26.0",
+    "@typescript-eslint/eslint-plugin": "~4.31.2",
+    "@typescript-eslint/parser": "~4.31.2",
+    "eslint": "^7.32.0",
     "jasmine-core": "~3.7.0",
     "karma": "~6.3.0",
     "karma-chrome-launcher": "~3.1.0",
@@ -58,6 +59,7 @@
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
     "ng-packagr": "^12.0.0",
+    "symlink-manager": "^1.5.0",
     "typescript": "~4.2.3"
   }
 }
