@@ -1,5 +1,4 @@
 ﻿using Microsoft.EntityFrameworkCore;
-using System;
 using Volo.Abp.EntityFrameworkCore.Modeling;
 
 namespace Volo.Abp.BlobStoring.Database.EntityFrameworkCore
@@ -7,21 +6,13 @@ namespace Volo.Abp.BlobStoring.Database.EntityFrameworkCore
     public static class BlobStoringDbContextModelCreatingExtensions
     {
         public static void ConfigureBlobStoring(
-            this ModelBuilder builder,
-            Action<BlobStoringModelBuilderConfigurationOptions> optionsAction = null)
+            this ModelBuilder builder)
         {
             Check.NotNull(builder, nameof(builder));
 
-            var options = new BlobStoringModelBuilderConfigurationOptions(
-                BlobStoringDatabaseDbProperties.DbTablePrefix,
-                BlobStoringDatabaseDbProperties.DbSchema
-            );
-
-            optionsAction?.Invoke(options);
-
             builder.Entity<DatabaseBlobContainer>(b =>
             {
-                b.ToTable(options.TablePrefix + "BlobContainers", options.Schema);
+                b.ToTable(BlobStoringDatabaseDbProperties.DbTablePrefix + "BlobContainers", BlobStoringDatabaseDbProperties.DbSchema);
 
                 b.ConfigureByConvention();
 
@@ -36,7 +27,7 @@ namespace Volo.Abp.BlobStoring.Database.EntityFrameworkCore
 
             builder.Entity<DatabaseBlob>(b =>
             {
-                b.ToTable(options.TablePrefix + "Blobs", options.Schema);
+                b.ToTable(BlobStoringDatabaseDbProperties.DbTablePrefix + "Blobs", BlobStoringDatabaseDbProperties.DbSchema);
 
                 b.ConfigureByConvention();
 
