@@ -1,5 +1,4 @@
-﻿using System;
-using Volo.Abp;
+﻿using Volo.Abp;
 using Volo.Abp.MongoDB;
 using Volo.Blogging.Blogs;
 using Volo.Blogging.Comments;
@@ -11,40 +10,33 @@ namespace Volo.Blogging.MongoDB
     public static class BloggingMongoDbContextExtensions
     {
         public static void ConfigureBlogging(
-            this IMongoModelBuilder builder,
-            Action<AbpMongoModelBuilderConfigurationOptions> optionsAction = null)
+            this IMongoModelBuilder builder)
         {
             Check.NotNull(builder, nameof(builder));
 
-            var options = new BloggingMongoModelBuilderConfigurationOptions(
-                BloggingDbProperties.DbTablePrefix
-            );
-
-            optionsAction?.Invoke(options);
-
             builder.Entity<BlogUser>(b =>
             {
-                b.CollectionName = options.CollectionPrefix + "Users";
+                b.CollectionName = BloggingDbProperties.DbTablePrefix + "Users";
             });
 
             builder.Entity<Blog>(b =>
             {
-                b.CollectionName = options.CollectionPrefix + "Blogs";
+                b.CollectionName = BloggingDbProperties.DbTablePrefix + "Blogs";
             });
 
             builder.Entity<Post>(b =>
             {
-                b.CollectionName = options.CollectionPrefix + "Posts";
+                b.CollectionName = BloggingDbProperties.DbTablePrefix + "Posts";
             });
 
             builder.Entity<Tagging.Tag>(b =>
             {
-                b.CollectionName = options.CollectionPrefix + "Tags";
+                b.CollectionName = BloggingDbProperties.DbTablePrefix + "Tags";
             });
 
             builder.Entity<Comment>(b =>
             {
-                b.CollectionName = options.CollectionPrefix + "Comments";
+                b.CollectionName = BloggingDbProperties.DbTablePrefix + "Comments";
             });
         }
     }
