@@ -1,5 +1,4 @@
-﻿using System;
-using Volo.Abp.IdentityServer.ApiResources;
+﻿using Volo.Abp.IdentityServer.ApiResources;
 using Volo.Abp.IdentityServer.ApiScopes;
 using Volo.Abp.IdentityServer.Clients;
 using Volo.Abp.IdentityServer.Devices;
@@ -12,45 +11,38 @@ namespace Volo.Abp.IdentityServer.MongoDB
     public static class AbpIdentityServerMongoDbContextExtensions
     {
         public static void ConfigureIdentityServer(
-            this IMongoModelBuilder builder,
-            Action<IdentityServerMongoModelBuilderConfigurationOptions> optionsAction = null)
+            this IMongoModelBuilder builder)
         {
             Check.NotNull(builder, nameof(builder));
 
-            var options = new IdentityServerMongoModelBuilderConfigurationOptions(
-                AbpIdentityServerDbProperties.DbTablePrefix
-            );
-
-            optionsAction?.Invoke(options);
-
             builder.Entity<ApiResource>(b =>
             {
-                b.CollectionName = options.CollectionPrefix + "ApiResources";
+                b.CollectionName = AbpIdentityServerDbProperties.DbTablePrefix + "ApiResources";
             });
 
             builder.Entity<ApiScope>(b =>
             {
-                b.CollectionName = options.CollectionPrefix + "ApiScopes";
+                b.CollectionName = AbpIdentityServerDbProperties.DbTablePrefix + "ApiScopes";
             });
 
             builder.Entity<IdentityResource>(b =>
             {
-                b.CollectionName = options.CollectionPrefix + "IdentityResources";
+                b.CollectionName = AbpIdentityServerDbProperties.DbTablePrefix + "IdentityResources";
             });
 
             builder.Entity<Client>(b =>
             {
-                b.CollectionName = options.CollectionPrefix + "Clients";
+                b.CollectionName = AbpIdentityServerDbProperties.DbTablePrefix + "Clients";
             });
 
             builder.Entity<PersistedGrant>(b =>
             {
-                b.CollectionName = options.CollectionPrefix + "PersistedGrants";
+                b.CollectionName = AbpIdentityServerDbProperties.DbTablePrefix + "PersistedGrants";
             });
 
             builder.Entity<DeviceFlowCodes>(b =>
             {
-                b.CollectionName = options.CollectionPrefix + "DeviceFlowCodes";
+                b.CollectionName = AbpIdentityServerDbProperties.DbTablePrefix + "DeviceFlowCodes";
             });
         }
     }
