@@ -1,5 +1,4 @@
-﻿using System;
-using JetBrains.Annotations;
+﻿using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore;
 using Volo.Abp.EntityFrameworkCore.Modeling;
 
@@ -8,21 +7,13 @@ namespace Volo.Abp.PermissionManagement.EntityFrameworkCore
     public static class AbpPermissionManagementDbContextModelBuilderExtensions
     {
         public static void ConfigurePermissionManagement(
-            [NotNull] this ModelBuilder builder,
-            [CanBeNull] Action<AbpPermissionManagementModelBuilderConfigurationOptions> optionsAction = null)
+            [NotNull] this ModelBuilder builder)
         {
             Check.NotNull(builder, nameof(builder));
 
-            var options = new AbpPermissionManagementModelBuilderConfigurationOptions(
-                AbpPermissionManagementDbProperties.DbTablePrefix,
-                AbpPermissionManagementDbProperties.DbSchema
-            );
-
-            optionsAction?.Invoke(options);
-
             builder.Entity<PermissionGrant>(b =>
             {
-                b.ToTable(options.TablePrefix + "PermissionGrants", options.Schema);
+                b.ToTable(AbpPermissionManagementDbProperties.DbTablePrefix + "PermissionGrants", AbpPermissionManagementDbProperties.DbSchema);
 
                 b.ConfigureByConvention();
 
