@@ -24,7 +24,7 @@
                     <abp-tab title="@group.DisplayName" name="v-pills-tab-@group.GetNormalizedGroupName()">
                         <h4>@group.DisplayName</h4>
                         <hr class="mt-2 mb-3"/>
-                        <div class="custom-scroll-content">
+                        <div class="w-100" style="max-height: 640px;overflow-y: auto">
                             <div class="pl-1 pt-1">
                                 <abp-input asp-for="@group.IsAllPermissionsGranted"
                                            check-box-hidden-input-render-mode="None"
