@@ -1,4 +1,3 @@
-using System;
 using Volo.Abp.MongoDB;
 
 namespace Volo.Abp.SettingManagement.MongoDB
@@ -6,20 +5,13 @@ namespace Volo.Abp.SettingManagement.MongoDB
     public static class SettingManagementMongoDbContextExtensions
     {
         public static void ConfigureSettingManagement(
-            this IMongoModelBuilder builder,
-            Action<SettingManagementMongoModelBuilderConfigurationOptions> optionsAction = null)
+            this IMongoModelBuilder builder)
         {
             Check.NotNull(builder, nameof(builder));
 
-            var options = new SettingManagementMongoModelBuilderConfigurationOptions(
-                AbpSettingManagementDbProperties.DbTablePrefix
-            );
-
-            optionsAction?.Invoke(options);
-
             builder.Entity<Setting>(b =>
             {
-                b.CollectionName = options.CollectionPrefix + "Settings";
+                b.CollectionName = AbpSettingManagementDbProperties.DbTablePrefix + "Settings";
             });
         }
     }
