@@ -1,4 +1,3 @@
-using System;
 using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore;
 using Volo.Abp.EntityFrameworkCore.Modeling;
@@ -8,8 +7,7 @@ namespace Volo.Abp.TenantManagement.EntityFrameworkCore
     public static class AbpTenantManagementDbContextModelCreatingExtensions
     {
         public static void ConfigureTenantManagement(
-            this ModelBuilder builder,
-            [CanBeNull] Action<AbpTenantManagementModelBuilderConfigurationOptions> optionsAction = null)
+            this ModelBuilder builder)
         {
             Check.NotNull(builder, nameof(builder));
 
@@ -18,16 +16,9 @@ namespace Volo.Abp.TenantManagement.EntityFrameworkCore
                 return;
             }
 
-            var options = new AbpTenantManagementModelBuilderConfigurationOptions(
-                AbpTenantManagementDbProperties.DbTablePrefix,
-                AbpTenantManagementDbProperties.DbSchema
-            );
-
-            optionsAction?.Invoke(options);
-
             builder.Entity<Tenant>(b =>
             {
-                b.ToTable(options.TablePrefix + "Tenants", options.Schema);
+                b.ToTable(AbpTenantManagementDbProperties.DbTablePrefix + "Tenants", AbpTenantManagementDbProperties.DbSchema);
 
                 b.ConfigureByConvention();
 
@@ -42,7 +33,7 @@ namespace Volo.Abp.TenantManagement.EntityFrameworkCore
 
             builder.Entity<TenantConnectionString>(b =>
             {
-                b.ToTable(options.TablePrefix + "TenantConnectionStrings", options.Schema);
+                b.ToTable(AbpTenantManagementDbProperties.DbTablePrefix + "TenantConnectionStrings", AbpTenantManagementDbProperties.DbSchema);
 
                 b.ConfigureByConvention();
 
