@@ -16,7 +16,6 @@ import { InitDirective } from './directives/init.directive';
 import { PermissionDirective } from './directives/permission.directive';
 import { ReplaceableTemplateDirective } from './directives/replaceable-template.directive';
 import { StopPropagationDirective } from './directives/stop-propagation.directive';
-import { VisibilityDirective } from './directives/visibility.directive';
 import { OAuthConfigurationHandler } from './handlers/oauth-configuration.handler';
 import { RoutesHandler } from './handlers/routes.handler';
 import { ApiInterceptor } from './interceptors/api.interceptor';
@@ -24,6 +23,7 @@ import { LocalizationModule } from './localization.module';
 import { ABP } from './models/common';
 import { LocalizationPipe } from './pipes/localization.pipe';
 import { SortPipe } from './pipes/sort.pipe';
+import { CookieLanguageProvider } from './providers/cookie-language.provider';
 import { LocaleProvider } from './providers/locale.provider';
 import { LocalizationService } from './services/localization.service';
 import { oAuthStorage } from './strategies/auth-flow.strategy';
@@ -64,7 +64,6 @@ export function storageFactory(): OAuthStorage {
     RouterOutletComponent,
     SortPipe,
     StopPropagationDirective,
-    VisibilityDirective,
   ],
   imports: [
     OAuthModule,
@@ -89,7 +88,6 @@ export function storageFactory(): OAuthStorage {
     RouterOutletComponent,
     SortPipe,
     StopPropagationDirective,
-    VisibilityDirective,
   ],
   providers: [LocalizationPipe],
   entryComponents: [
@@ -131,6 +129,7 @@ export class CoreModule {
       ngModule: RootCoreModule,
       providers: [
         LocaleProvider,
+        CookieLanguageProvider,
         {
           provide: 'CORE_OPTIONS',
           useValue: options,
