@@ -1,6 +1,7 @@
+export * from './app-config.token';
+export * from './cookie-language-key.token';
 export * from './list.token';
 export * from './lodaer-delay.token';
 export * from './manage-profile.token';
 export * from './options.token';
-export * from './app-config.token';
 export * from './tenant-key.token';
