@@ -1,5 +1,4 @@
+export * from './ellipsis.directive';
 export * from './loading.directive';
 export * from './ngx-datatable-default.directive';
 export * from './ngx-datatable-list.directive';
-export * from './table-sort.directive';
-export * from './ellipsis.directive';
