@@ -1,5 +1,4 @@
 export * from './confirmation.service';
-export * from './modal.service';
-export * from './toaster.service';
 export * from './nav-items.service';
 export * from './page-alert.service';
+export * from './toaster.service';
