@@ -17,11 +17,7 @@ import { HttpErrorWrapperComponent } from './components/http-error-wrapper/http-
 import { LoaderBarComponent } from './components/loader-bar/loader-bar.component';
 import { LoadingComponent } from './components/loading/loading.component';
 import { ModalCloseDirective } from './components/modal/modal-close.directive';
-import { ModalContainerComponent } from './components/modal/modal-container.component';
 import { ModalComponent } from './components/modal/modal.component';
-import { SortOrderIconComponent } from './components/sort-order-icon/sort-order-icon.component';
-import { TableEmptyMessageComponent } from './components/table-empty-message/table-empty-message.component';
-import { TableComponent } from './components/table/table.component';
 import { ToastContainerComponent } from './components/toast-container/toast-container.component';
 import { ToastComponent } from './components/toast/toast.component';
 import { DEFAULT_VALIDATION_BLUEPRINTS } from './constants/validation';
@@ -29,7 +25,6 @@ import { EllipsisModule } from './directives/ellipsis.directive';
 import { LoadingDirective } from './directives/loading.directive';
 import { NgxDatatableDefaultDirective } from './directives/ngx-datatable-default.directive';
 import { NgxDatatableListDirective } from './directives/ngx-datatable-list.directive';
-import { TableSortDirective } from './directives/table-sort.directive';
 import { ErrorHandler } from './handlers/error.handler';
 import { initLazyStyleHandler } from './handlers/lazy-style.handler';
 import { RootParams } from './models/common';
@@ -46,15 +41,11 @@ const declarationsWithExports = [
   LoaderBarComponent,
   LoadingComponent,
   ModalComponent,
-  TableComponent,
-  TableEmptyMessageComponent,
   ToastComponent,
   ToastContainerComponent,
-  SortOrderIconComponent,
   NgxDatatableDefaultDirective,
   NgxDatatableListDirective,
   LoadingDirective,
-  TableSortDirective,
   ModalCloseDirective,
 ];
 
@@ -66,13 +57,12 @@ const declarationsWithExports = [
     NgbPaginationModule,
     EllipsisModule,
   ],
-  declarations: [...declarationsWithExports, HttpErrorWrapperComponent, ModalContainerComponent],
+  declarations: [...declarationsWithExports, HttpErrorWrapperComponent],
   exports: [NgxDatatableModule, EllipsisModule, ...declarationsWithExports],
   providers: [DatePipe],
   entryComponents: [
     HttpErrorWrapperComponent,
     LoadingComponent,
-    ModalContainerComponent,
     ToastContainerComponent,
     ConfirmationComponent,
   ],
