@@ -4,8 +4,8 @@ const childProcess = require('child_process');
 const { program } = require('commander');
 
 program.version('0.0.1');
-program.option('-r, --rc', 'whether version is rc');
-program.option('-rg, --registry <registry>', 'target npm server registry')
+program.option('-pr, --prerelase', 'whether version is prerelase');
+program.option('-rg, --registry <registry>', 'target npm server registry');
 program.parse(process.argv);
 
 const packages = (process.argv[3] || 'abp').split(',').join('|');
@@ -15,7 +15,7 @@ const check = (pkgJsonPath) => {
     return childProcess
       .execSync(
         `ncu "/^@(${packages}).*$/" --packageFile ${pkgJsonPath} -u${
-          program.rc ? ' --target greatest' : ''
+          program.prerelase ? ' --target greatest' : ''
         }${program.registry ? ` --registry ${program.registry}` : ''}`
       )
       .toString();
