@@ -1,16 +1,18 @@
 const { program } = require('commander');
 const fse = require('fs-extra');
+const semverParse = require('semver/functions/parse');
 
 program.version('0.0.1');
 program.option('-n, --nextVersion', 'version in common.props');
-program.option('-r, --rc', 'whether version is rc');
+program.option('-pr, --prerelase', 'whether version is prerelase');
 program.option('-cv, --customVersion <customVersion>', 'set exact version');
 
 program.parse(process.argv);
 
 if (program.nextVersion) console.log(getVersion());
 
-if (program.rc) console.log(getVersion().includes('rc'));
+if (program.prerelase)
+  console.log(!!semverParse(getVersion()).prerelease?.length);
 
 function getVersion() {
   if (program.customVersion) return program.customVersion;
