@@ -1,5 +1,4 @@
-﻿using System;
-using Microsoft.EntityFrameworkCore;
+﻿using Microsoft.EntityFrameworkCore;
 using Volo.Abp;
 
 namespace MyCompanyName.MyProjectName.EntityFrameworkCore
@@ -7,30 +6,22 @@ namespace MyCompanyName.MyProjectName.EntityFrameworkCore
     public static class MyProjectNameDbContextModelCreatingExtensions
     {
         public static void ConfigureMyProjectName(
-            this ModelBuilder builder,
-            Action<MyProjectNameModelBuilderConfigurationOptions> optionsAction = null)
+            this ModelBuilder builder)
         {
             Check.NotNull(builder, nameof(builder));
-
-            var options = new MyProjectNameModelBuilderConfigurationOptions(
-                MyProjectNameDbProperties.DbTablePrefix,
-                MyProjectNameDbProperties.DbSchema
-            );
-
-            optionsAction?.Invoke(options);
 
             /* Configure all entities here. Example:
 
             builder.Entity<Question>(b =>
             {
                 //Configure table & schema name
-                b.ToTable(options.TablePrefix + "Questions", options.Schema);
-            
+                b.ToTable(MyProjectNameDbProperties.DbTablePrefix + "Questions", MyProjectNameDbProperties.DbSchema);
+
                 b.ConfigureByConvention();
-            
+
                 //Properties
                 b.Property(q => q.Title).IsRequired().HasMaxLength(QuestionConsts.MaxTitleLength);
-                
+
                 //Relations
                 b.HasMany(question => question.Tags).WithOne().HasForeignKey(qt => qt.QuestionId);
 
