@@ -1,5 +1,4 @@
-﻿using System;
-using Volo.Abp;
+﻿using Volo.Abp;
 using Volo.Abp.MongoDB;
 
 namespace MyCompanyName.MyProjectName.MongoDB
@@ -7,16 +6,9 @@ namespace MyCompanyName.MyProjectName.MongoDB
     public static class MyProjectNameMongoDbContextExtensions
     {
         public static void ConfigureMyProjectName(
-            this IMongoModelBuilder builder,
-            Action<AbpMongoModelBuilderConfigurationOptions> optionsAction = null)
+            this IMongoModelBuilder builder)
         {
             Check.NotNull(builder, nameof(builder));
-
-            var options = new MyProjectNameMongoModelBuilderConfigurationOptions(
-                MyProjectNameDbProperties.DbTablePrefix
-            );
-
-            optionsAction?.Invoke(options);
         }
     }
 }
